+++ src/peer_manager/tests.rs
@@ -217,7 +217,6 @@ fn make_dynamic_manager_config() -> Config {
                 log_format: "json".to_string(),
                 grpc_tcp: None,
                 grpc_uds: None,
-                looking_glass: None,
             },
             dynamic_neighbor_limit: Some(100),
             worker_threads: None,